Move tests into a separate folder
