+++ src/tests/ChangeEmailPage.test.js
@@ -1,17 +1,19 @@
 import React from 'react';
-import SignInForm from './SignInForm';
+import ChangeEmailPage from '../components/ChangeEmailPage';
 import {MemoryRouter} from 'react-router';
 import {mount} from 'enzyme';
 
 test('should display page layout', () => {
-  const initialEntries = ["/sign-in-form"];
+  const initialEntries = ["/change-email-page"];
   const initialIndex = 0;
   const wrapper = mount(
     <MemoryRouter
       initialIndex={initialIndex}
       initialEntries={initialEntries}>
-        <SignInForm/>
+        <ChangeEmailPage/>
     </MemoryRouter>
   );
-  expect(wrapper.find('input').length).toBe(2);
+  expect(wrapper.find('label').length).toBe(1)
+  expect(wrapper.find('label').text()).toBe("Enter new Email")
+  expect(wrapper.find('input').length).toBe(1)
 });
