+++ src/tests/ChangePasswordPage.test.js
@@ -1,19 +1,19 @@
 import React from 'react';
-import ChangeEmailPage from './ChangeEmailPage';
+import ChangePasswordPage from '../components/ChangePasswordPage';
 import {MemoryRouter} from 'react-router';
 import {mount} from 'enzyme';
 
 test('should display page layout', () => {
-  const initialEntries = ["/change-email-page"];
+  const initialEntries = ["/change-password-page"];
   const initialIndex = 0;
   const wrapper = mount(
     <MemoryRouter
       initialIndex={initialIndex}
       initialEntries={initialEntries}>
-        <ChangeEmailPage/>
+        <ChangePasswordPage/>
     </MemoryRouter>
   );
   expect(wrapper.find('label').length).toBe(1)
-  expect(wrapper.find('label').text()).toBe("Enter new Email")
+  expect(wrapper.find('label').text()).toBe("Enter new password")
   expect(wrapper.find('input').length).toBe(1)
 });
