+++ src/tests/EditProfilePage.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import EditProfilePage from './EditProfilePage';
+import EditProfilePage from '../components/EditProfilePage';
 import {MemoryRouter} from 'react-router';
 import {mount} from 'enzyme';
