+++ src/tests/ProfilePage.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProfilePage from './ProfilePage';
+import ProfilePage from '../components/ProfilePage';
 import {MemoryRouter} from 'react-router';
 import {mount} from 'enzyme';
