+++ src/tests/SignInForm.test.js
@@ -1,17 +1,17 @@
 import React from 'react';
-import CreateAccountForm from './CreateAccountForm';
+import SignInForm from '../components/SignInForm';
 import {MemoryRouter} from 'react-router';
 import {mount} from 'enzyme';
 
 test('should display page layout', () => {
-  const initialEntries = ["/create-account"];
+  const initialEntries = ["/sign-in-form"];
   const initialIndex = 0;
   const wrapper = mount(
     <MemoryRouter
       initialIndex={initialIndex}
       initialEntries={initialEntries}>
-        <CreateAccountForm/>
+        <SignInForm/>
     </MemoryRouter>
   );
-  expect(wrapper.find('input').length).toBe(4)
+  expect(wrapper.find('input').length).toBe(2);
 });
